### 2022-05-02 #4271 @Viech
This looks cool but I won't have time to review this month due to holidays and an official trip. If this gets stuck in june, I can fill in.
