### 2022-05-05 #4271 @stsievert
@diehlpk I changed my mind. I can do it, but it won't be prompt or even quick.
