### 2022-05-05 #4377 @editorialbot
[REVIEW]: PyCUTEst: an open source Python package of optimization test problems
**Submitting author:** @jfowkes (<a href="http://orcid.org/0000-0002-8048-4572">Jaroslav Fowkes</a>)
**Repository:** https://github.com/jfowkes/pycutest
**Branch with paper.md** (empty if default branch):
**Version:** v1.4
