### 2022-06-20 #4377 @stsievert
I have checked (or have items below addressing) all the boxes in https://github.com/openjournals/joss-reviews/issues/4377#issuecomment-1118435375. I do not think this paper is ready for acceptance into the JOSS, at least as-is. Some clearly blocking issues:

* I did not find any guidelines for third parties wishing to contribute, report and find support.
* Who's Arpad Burmen, and why is he listed as an author? He's listed as a first author, but [doesn't have any commits][commits] to this project.

I had significant difficulty installing this software on macOS 10.14 to complete this review; I had to put PyCUTEst in it's own Dockerfile and complete my review there. I think the install docs could use some improvement/modification. Here's my Dockerfile:

<summary>The Dockerfile I used to install PyCUTEst</summary>

```
FROM continuumio/miniconda3
# docker build -t pyc .  # build the container
# docker run -dt pyc  # launch the container
# docker exec -it f651b5e17dec /bin/bash  # log in to the container

WORKDIR /pycutest
RUN apt update
RUN apt install -y build-essential git gfortran
RUN git clone https://github.com/ralna/ARCHDefs ./archdefs
RUN git clone https://github.com/ralna/SIFDecode ./sifdecode
RUN git clone https://github.com/ralna/CUTEst ./cutest
RUN git clone https://bitbucket.org/optrove/sif ./mastsif

ENV ARCHDEFS /pycutest/archdefs/
ENV SIFDECODE /pycutest/sifdecode/
ENV MASTSIF /pycutest/mastsif/
ENV CUTEST /pycutest/cutest/
ENV MYARCH "pc64.lnx.gfo"
ENTRYPOINT ./archdefs/install_optrove && tail -f /dev/null
```

Other issues/style nits include the following (in no particular order):

* I'm not sure I'd say the documentation [homepage] "clearly state[s]" the problem of interest, certainly not as clear as the paper.
* Types are missing on the API documentation (e.g., what's a "Lagrange multiplies" in [CUTEstProblem.hess][hess], and what's the shape of it? Is the "input vector" a NumPy ndarray, or can it be a PyTorch Tensor?)
* The [testing section of the documentation][testing] is not complete, or misnamed.
* I tried to install with pip. It didn't work, I got a RuntimeError: "Could not find CUTEST installation - have CUTEST and MYARCH environment variables been set correctly?"
    * I first installed with pip, then I tried homebrew (which hung), then I (successfully) tried a Dockerfile.
* There are multiple files in `examples/` with no clear purpose. e.g.,
    * [allintic.py] only defines functions. What's the purpose?
    * [param_warning_test.py] raises a RuntimeError "SIFDECODE error:  Failed setting HNS to 4". (which looks expected, and [is caught in the tests]).
    * I get errors on [test_cutest_full.py], [test_dimensions.py] and [test_cutest_ls.py], I think because new breaking versions have been released (no attribute `getinfo`).
    * [test_available_params_print.py] doesn't do much (only prints available parameters).
    * (note: successful/proper examples include [run_demo_problem.py], [solve_rosenbrock.py])

[homepage]:https://jfowkes.github.io/pycutest/
[is caught in the tests]:https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/pycutest/tests/test_basic_functionality.py#L78
[solve_rosenbrock.py]:https://github.com/jfowkes/pycutest/blob/master/examples/solve_rosenbrock.py
[run_demo_problem.py]:https://github.com/jfowkes/pycutest/blob/master/examples/run_demo_problem.py
[test_dimensions.py]:https://github.com/jfowkes/pycutest/blob/master/examples/test_dimensions.py
[test_cutest_ls.py]:https://github.com/jfowkes/pycutest/blob/master/examples/test_cutest_ls.py
[test_available_params_print.py]:https://github.com/jfowkes/pycutest/blob/master/examples/test_available_params_print.py
[test_cutest_full.py]:https://github.com/jfowkes/pycutest/blob/master/examples/test_cutest_full.py
[param_warning_test.py]:https://github.com/jfowkes/pycutest/blob/master/examples/param_warning_test.py
[allintic.py]:https://github.com/jfowkes/pycutest/blob/master/examples/allinitc.py

[hess]:https://jfowkes.github.io/pycutest/_build/html/methods/pycutest.CUTEstProblem.hess.html
[commits]:https://github.com/jfowkes/pycutest/graphs/contributors?from=2018-05-27&to=2022-06-20&type=a
[testing]:https://jfowkes.github.io/pycutest/_build/html/install.html#testing
