### 2022-06-27 #4377 @jfowkes
@stsievert thank you very much for taking the time to review our submission, much appreciated. In response to the two blocking issues you raised:
- We will of course add guidelines for contributors to the readme and docs, apologies for the omission, we missed that this was a requirement.
- As stated at the start of the readme and docs, Arpad Burmen wrote the original PyCUTEr of which PyCUTEst is an evolution. He is responsible for writing the majority of `c_interface.py` (as acknowledged in the code headers) and is thus a major contributor. Arpad Burmen has no commits to the project as his PyCUTEr was never under version control and dates back to 2011 (please see his PyCUTEr website).

I'm sorry to hear that you had difficulties installing on macOS but version 10.14 is rather outdated now and no longer supported by homebrew. Our CI tests PyCUTEst installation with homebrew on the latest macOS as well as Linux (see our Github Actions). We are of course happy to add your dockerfile as an alternative installation option to the docs.

Regarding the other issues you raised:
- We agree the homepage could state the problem of interest more clearly, we will update it to more closely align with the problem statement in the paper.
- We agree that the API documentation should have types, we will add these.
- We agree that the testing section of the documentation is misnamed, this is a holdover from when we used a setup.py entry-point for tests.
- As the docs state, CUTEst is required for installation (we are a python fronted for it)
- Regarding the issues with the tests, we will carefully look into these: it is possible that new breaking versions have been released in the past few months (all of the tests were passing on our CI as of our last release). We really should have a scheduled GitHub Action that regularly runs our CI tests, we will add this in due course.

@diehlpk what is the procedure going forward? Do we make the requested changes and then come back here for a re-review?
