### 2022-06-27 a69cf1d
Fix typo in README
