### 2022-06-28 186b736
Add types to problem method docstrings
