+++ pycutest/problem_class.py
@@ -256,7 +256,9 @@ class CUTEstProblem(object):
         This calls CUTEst routine CUTEst_cfn.
 
         :param x: input vector
+        :type x: numpy.ndarray with shape (n,)
         :return: tuple (f, c) of objective and constraint values at x for constrained problems.
+        :rtype: (float, numpy.ndarray(m,))
         """
         self.check_input_x(x)
         f, c = self._module.objcons(self.free_to_all(x))
@@ -279,8 +281,11 @@ class CUTEstProblem(object):
         This calls CUTEst routine CUTEST_uofg or CUTEST_cofg.
 
         :param x: input vector
+        :type x: numpy.ndarray with shape (n,)
         :param gradient: whether to return objective and gradient, or just objective (default=False; i.e. objective only)
+        :type gradient: bool, optional
         :return: objective value f, or tuple (f,g) of objective and gradient at x
+        :rtype: float or (float, numpy.ndarray(n,))
         """
         self.check_input_x(x)
         if gradient:
@@ -294,7 +299,7 @@ class CUTEstProblem(object):
 
     def cons(self, x, index=None, gradient=False):
         """
-        Evaluate the constraints (and optionally their Jacobian).
+        Evaluate the constraints (and optionally their Jacobian or gradient).
 
         .. code-block:: python
 
@@ -314,9 +319,13 @@ class CUTEstProblem(object):
         For large problems, problem.scons returns sparse matrices.
 
         :param x: input vector
+        :type x: numpy.ndarray with shape (n,)
         :param index: which constraint to evaluate (default=None -> all constraints). Must be in 0..self.m-1.
+        :type index: int, optional
         :param gradient: whether to return constraint(s) and gradient/Jacobian, or just constraint (default=False; i.e. constraint only)
-        :return: value of constraint(s) and Jacobian/gradient of constraint(s) at x
+        :type gradient: bool, optional
+        :return: value of constraint(s), and Jacobian or gradient of constraint(s) at x
+        :rtype: numpy.ndarray(m,) or float or (numpy.ndarray(m,), numpy.ndarray(m,n)) or (float, numpy.ndarray(n,))
         """
         if self.m <= 0:
             return None
@@ -342,7 +351,7 @@ class CUTEstProblem(object):
 
     def lagjac(self, x, v=None):
         """
-        Evaluate gradient of objective/Lagrangian, and Jacobian of constraints.
+        Evaluate gradient of objective or Lagrangian, and Jacobian of constraints.
 
         .. code-block:: python
 
@@ -360,8 +369,11 @@ class CUTEstProblem(object):
         Note: in CUTEst, the sign convention is such that the Lagrangian = objective + lagrange_multipliers * constraints
 
         :param x: input vector
+        :type x: numpy.ndarray with shape (n,)
         :param v: input vector of Lagrange multipliers
-        :return: gradient of objective/Lagrangian, and Jacobian of constraints
+        :type v: numpy.ndarray with shape (m,), optional
+        :return: gradient of objective or Lagrangian, and Jacobian of constraints
+        :rtype: (numpy.ndarray(n,), numpy.ndarray(m,n))
         """
         self.check_input_x(x)
         if v is None:
@@ -380,9 +392,9 @@ class CUTEstProblem(object):
 
         .. code-block:: python
 
-            # evaluate J*p where J is the last compute Jacobian
+            # evaluate J*p where J is the last computed Jacobian
             r = problem.jprod(p)
-            # evaluate J.T*p where J is the last compute Jacobian
+            # evaluate J.T*p where J is the last computed Jacobian
             r = problem.jprod(p, transpose=True)
[…]
