### 2022-06-28 c4a9cbc
Remove testing section from docs
