### 2022-06-28 e9bd42a
Add scheduled CI tests
