+++ setup.py
@@ -35,7 +35,12 @@ setup(
     author='Jaroslav Fowkes and Lindon Roberts',
     author_email='jaroslav.fowkes@maths.ox.ac.uk',
     url="https://github.com/jfowkes/pycutest/",
-    download_url="https://github.com/jfowkes/pycutest/archive/v1.2.tar.gz",
+    download_url="https://github.com/jfowkes/pycutest/releases/",
+    project_urls={
+        "Bug Tracker": "https://github.com/jfowkes/pycutest/issues/",
+        "Documentation": "https://jfowkes.github.io/pycutest/",
+        "Source Code": "https://github.com/jfowkes/pycutest/",
+    },
     packages=['pycutest'],
     license='GNU GPL',
     keywords = "mathematics optimization",
