### 2022-06-28 d486551
Add dockerfile as alternate install option
