+++ docs/install.rst
@@ -1,5 +1,5 @@
-Installing PyCUTEst
-===================
+Installation
+============
 
 Requirements
 ------------
@@ -15,8 +15,12 @@ Additionally, the following python packages should be installed (these will be i
 
 **Please Note:** Currently PyCUTEst only supports Mac and Linux. For Windows 10 (or later), PyCUTEst can be used through the `Windows Subsystem for Linux <https://docs.microsoft.com/en-us/windows/wsl/>`_, following the Linux installation instructions.
 
+Installing CUTEst
+-----------------
+Note that CUTEst must be installed in order for PyCUTEst to work.
+
 Installing CUTEst on Linux
---------------------------
+^^^^^^^^^^^^^^^^^^^^^^^^^^
 These instructions do not include installation of the MATLAB interface. You will need to install four packages: `archdefs <https://github.com/ralna/ARCHDefs>`_, `SIFDecode <https://github.com/ralna/SIFDecode>`_, `CUTEst <https://github.com/ralna/CUTEst>`_ and `MASTSIF <https://bitbucket.org/optrove/sif>`_. To keep things simple, install all four packages in the same directory:
 
  .. code-block:: bash
@@ -73,7 +77,7 @@ And CUTEst should run from here. To test that the installation works, issue the 
 **Please Note:** *currently PyCUTEst only supports gfortran and uses the default version on your path (as returned by* :code:`gfortran -v` *). Please ensure this is the same version that you install CUTEst with above otherwise you may experience segmentation faults, this should be the case if you select the generic* :code:`GNU gfortran compiler` *as the fortran compiler in the installer above.*
 
 Installing CUTEst on Mac
-------------------------
+^^^^^^^^^^^^^^^^^^^^^^^^
 Install CUTEst using Homebrew as detailed below (installing CUTEst manually on Mac is not supported). First it is important to ensure that you have the latest version of Xcode Command Line Tools installed (or the latest version of Xcode), please ensure this is the case by following `this guide <https://mac.install.guide/commandlinetools/index.html>`_. Now install the Homebrew package manager:
 
  .. code-block:: bash
@@ -93,8 +97,45 @@ Then you can easily install CUTEst:
 
 **Please Note:** *you may see warnings such as* :code:`ld: warning: object file (RANGE.o) was built for newer macOS version (11.5) than being linked (10.15)` *when using PyCUTEst on Mac, to suppress these warnings please set the environment variable* :code:`MACOSX_DEPLOYMENT_TARGET` *to your current macOS version (e.g.* :code:`export MACOSX_DEPLOYMENT_TARGET=11.5` *in this example, you can make this permanent by adding it to your* :code:`~/.bashrc` *file).*
 
+CUTEst Docker Container
+^^^^^^^^^^^^^^^^^^^^^^^
+CUTEst can be installed into its own Docker container using the following Dockerfile:
+
+ .. code-block:: docker
+
+    FROM continuumio/miniconda3
+
+    WORKDIR /cutest
+    RUN apt update
+    RUN apt install -y build-essential git gfortran
+    RUN git clone https://github.com/ralna/ARCHDefs ./archdefs
+    RUN git clone https://github.com/ralna/SIFDecode ./sifdecode
+    RUN git clone https://github.com/ralna/CUTEst ./cutest
+    RUN git clone https://bitbucket.org/optrove/sif ./mastsif
+
+    ENV ARCHDEFS /cutest/archdefs/
+    ENV SIFDECODE /cutest/sifdecode/
+    ENV MASTSIF /cutest/mastsif/
+    ENV CUTEST /cutest/cutest/
+    ENV MYARCH "pc64.lnx.gfo"
[…]
