### 2022-06-28 0032383
Add dockerfile as alternate install option
