### 2022-06-29 #4377 @stsievert
> Regarding the issues with the tests, we will carefully look into these: it is possible that new breaking versions have been released in the past few months (all of the tests were passing on our CI as of our last release). We really should have a scheduled GitHub Action that regularly runs our CI tests, we will add this in due course.

'scuse me, I mean that some files *in the `examples/` directory* raise errors (e.g., [examples/test_cutest_full.py][ls]). All these files were added 4 years ago. The fact that two useful examples (e.g., [examples/solve_rosenbrock.py][sr]) still run successfully is rather impressive.

[sr]:https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/examples/solve_rosenbrock.py
[ls]:https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/examples/test_cutest_full.py
[f]:https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/examples/test_cutest_full.py
