### 2022-07-04 #4377 @jfowkes
> 'scuse me, I mean that some files _in the `examples/` directory_ raise errors (e.g., [examples/test_cutest_full.py](https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/examples/test_cutest_full.py)). All these files were added 4 years ago. The fact that two useful examples (e.g., [examples/solve_rosenbrock.py](https://github.com/jfowkes/pycutest/blob/886421736ea4aff0ce78212f8ec0494fa69bb8f8/examples/solve_rosenbrock.py)) still run successfully is rather impressive.

Ah I see, thank you. I will investigate and update these to run correctly.
