### 2022-07-04 f996057
Add contributor guidelines
