### 2022-07-04 246b7cc
Add contributor guidelines
