### 2022-07-04 c920d19
Remove ancient tests and add new example
