+++ examples/allinitc.py
@@ -1,98 +1,0 @@
-"""
-Nick's "all in it" test problem (constrained version)
-
-   Source:
-   N. Gould, private communication.
-
-   SIF input: Nick Gould, June 1990.
-
-   classification OUR2-AY-4-0
-"""
-# Ensure compatibility with Python 2
-from __future__ import absolute_import, division, print_function, unicode_literals
-
-import numpy as np
-import math as ma
-
-# objective
-def f(x):
-	return np.array([
-	x[2]-1 + x[0]**2 + x[1]**2 + (x[2]+x[3])**2 + ma.sin(x[2])**2 + x[0]**2*x[1]**2 + x[3]-3 +
-	ma.sin(x[2])**2 + (x[3]-1)**2 + (x[1]**2)**2 + (x[2]**2 + (x[3]+x[0])**2)**2 +
-	(x[0]-4 + ma.sin(x[3])**2 + x[1]**2*x[2]**2)**2 + ma.sin(x[3])**4
-	])
-
-# gradient
-def g(x):
-	return np.array([
-	2*(-4 + 2*x[0] + x[0]*x[1]**2 + x[1]**2*x[2]**2 + 2*(x[0]+x[3]) *
-	 (x[2]**2 + (x[0]+x[3])**2) + ma.sin(x[3])**2), 
-    2*x[1] * (1 + x[0]**2 + 2*x[1]**2 + 2*x[2]**2 * (-4 + x[0] + x[1]**2 * x[2]**2 + ma.sin(x[3])**2)), 
-    1 + 2*(x[2]+x[3]) + 4*x[2] * (x[2]**2 + (x[0]+x[3])**2) + 4*ma.cos(x[2])*ma.sin(x[2]) + 
-     4*x[1]**2*x[2] * (-4 + x[0] + x[1]**2*x[2]**2 + ma.sin(x[3])**2), 
-    -1 + 2*x[2] + 4*x[3] + 4*(x[0]+x[3])*(x[2]**2 + (x[0]+x[3])**2) + 
-     4*ma.cos(x[3])*ma.sin(x[3]) * (-4 + x[0] + x[1]**2*x[2]**2 + 2*ma.sin(x[3])**2)
-	])
-
-# Hessian
-def H(x):
-	return np.array([
-	[2*(2 + x[1]**2 + 2*x[2]**2 + 6*(x[0]+x[3])**2), 
-	 4*x[1]*(x[0]+x[2]**2), 
-     4*x[2]*(x[1]**2 + 2*(x[0]+x[3])), 
-     4*(x[2]**2 + 3*(x[0] + x[3])**2 + ma.cos(x[3])*ma.sin(x[3]))
-    ],
-    [4*x[1]*(x[0] + x[2]**2), 
-     2*(1 + x[0]**2 - 8*x[2]**2 + 2*x[0]*x[2]**2 + 6*x[1]**2*(1 + x[2]**4) + 2*x[2]**2*ma.sin(x[3])**2), 
-     4*x[1]*x[2]*(-7 + 2*x[0] + 4*x[1]**2*x[2]**2 - ma.cos(2*x[3])), 
-     8*x[1]*x[2]**2*ma.cos(x[3])*ma.sin(x[3])
-    ],
-    [4*x[2]*(x[1]**2 + 2*(x[0]+x[3])), 
-     4*x[1]*x[2]*(-7 + 2*x[0] + 4*x[1]**2*x[2]**2 - ma.cos(2*x[3])), 
-     2 + 12*x[2]**2 + 8*x[1]**4*x[2]**2 + 4*(x[0] + x[3])**2 + 4*ma.cos(x[2])**2 - 
[…]
