### 2022-07-04 6ea3d6d
Remove ancient tests and add new example
