+++ examples/solve_rosenbrock.py
@@ -1,7 +1,7 @@
 """
 PyCUTEst example: minimize 2D Rosenbrock function using Newton's method.
 
-Jaroslav Fowkes and Lindon Roberts, 2018.
+Jaroslav Fowkes and Lindon Roberts, 2022.
 """
 
 # Ensure compatibility with Python 2
