+++ docs/example.rst
@@ -1,6 +1,8 @@
 Example Usage
 =============
 
+General optimization
+--------------------
 The following code presents a simple use of PyCUTEst to minimize `Rosenbrock's function <https://en.wikipedia.org/wiki/Rosenbrock_function>`_ in 2D (problem :code:`ROSENBR`) using Newton's method.
 
 .. code-block:: python
@@ -8,7 +10,7 @@ The following code presents a simple use of PyCUTEst to minimize `Rosenbrock's f
     """
     PyCUTEst example: minimize 2D Rosenbrock function using Newton's method.
 
-    Jaroslav Fowkes and Lindon Roberts, 2020.
+    Jaroslav Fowkes and Lindon Roberts, 2022.
     """
 
     # Ensure compatibility with Python 2
@@ -52,3 +54,63 @@ For this problem, Newton's method finds the unique local and global minimum :cod
     Found minimum x = [1. 1.] after 7 iterations
     Done
 
+Nonlinear least-squares
+-----------------------
+In CUTEst nonlinear least-squares problems are encoded as nonlinear equations which have no objective function defined (as denoted by the :code:`N` in their classification).
+Instead, their residual(s) :code:`r(x)` and Jacobian :code:`J(x)` are accessed using `CUTEstProblem.cons() <methods/pycutest.CUTEstProblem.cons.html>`_.
+Typically you will want to use them as nonlinear least-squares problems, in which case you can construct a standard objective function and gradient as:
+
+.. code-block:: latex
+
+    f(x) = 0.5*||r(x)||^2
+    g(x) = J(x)^T r(x)
+
+The following code presents a simple use of PyCUTEst to minimize problem :code:`ARGLALE` using the `Gauss-Newton algorithm <https://en.wikipedia.org/wiki/Gauss–Newton_algorithm>`_.
+
+.. code-block:: python
+
+    """
+    PyCUTEst example: minimize 10D ALRGLALE problem using the Gauss-Newton algorithm.
+
+    Jaroslav Fowkes and Lindon Roberts, 2022.
+    """
+
+    # Ensure compatibility with Python 2
+    from __future__ import print_function
+    import numpy as np
+    import pycutest
+
+    # Nonlinear least-squares problem in 10 dimensions with 20 residuals
+    p = pycutest.import_problem('ARGLALE', sifParams={'N':10, 'M':20})
+
+    print("ARGLALE problem in %gD with %g residuals" % (p.n, p.m))
+
+    iters = 0
+
+    x = p.x0
+    r, J = p.cons(x, gradient=True)  # residual and Jacobian
+    f = 0.5 * np.dot(r, r)  # objective
+    g = J.T.dot(r)  # gradient
+    H = J.T.dot(J)  # Gauss-Newton Hessian approximation
+
+    while iters < 100 and np.linalg.norm(g) > 1e-10:
+        print("Iteration %g: objective value is %g with norm of gradient %g at x = %s" % (iters, f, np.linalg.norm(g), str(x)))
+        s = np.linalg.solve(H, -g)  # Gauss-Newton step
+        x = x + s  # used fixed step length
+        r, J = p.cons(x, gradient=True)
+        f = 0.5 * np.dot(r, r)
+        g = J.T.dot(r)
+        H = J.T.dot(J)
+        iters += 1
+
+    print("Found minimum x = %s after %g iterations" % (str(x), iters))
+    print("Done")
+
+For this problem, the Gauss-Newton algorithm finds the unique local and global minimum :code:`f(-1,-1,-1,-1,-1,-1,-1,-1,-1,-1)=5` in one step:
+
+.. code-block:: none
+
[…]
