### 2022-07-12 54e558f
Minor documentation edits
