### 2022-07-12 #4377 @jfowkes
@diehlpk @jonjoncardoso @stsievert we have now made all the requested changes, could you please re-review?

Please see the PR jfowkes/pycutest#23 for the full changelist (in particular I have removed the ancient test files from the `examples/` directory and added and documented another useful usage example `solve_arglale.py`).
