### 2022-08-06 #4377 @stsievert
I'm not seeing instructions for third parties on how to "report issues or problems with the software" or "seek support" in the documentation, though I am seeing [some guidelines][c] on how to "contribute to the software".

The author order isn't clear. Who's the first author? @jfowkes is listed as the first author on [the documentation][doc], and Arpad Buermen is listed as the first author on the paper (and isn't mentioned as an author of [the documentation][doc]). I think the first author should be the person primarily responsible for the PyCUTEst interface, tests and documentation (the focus of this paper), which may different from the person responsible for the underlying interface.

The Docker installation instructions didn't quite work for me; some more detail is below:

I used this Dockefile to clean the installation process for fewer questions:

``` Dockerfile
FROM continuumio/miniconda3

WORKDIR /cutest
RUN apt update
RUN apt install -y build-essential git gfortran
RUN git clone https://github.com/ralna/ARCHDefs ./archdefs
RUN git clone https://github.com/ralna/SIFDecode ./sifdecode
RUN git clone https://github.com/ralna/CUTEst ./cutest
RUN git clone https://bitbucket.org/optrove/sif ./mastsif

ENV ARCHDEFS /cutest/archdefs/
ENV SIFDECODE /cutest/sifdecode/
ENV MASTSIF /cutest/mastsif/
ENV CUTEST /cutest/cutest/
ENV MYARCH "pc64.lnx.gfo"

RUN echo "let install_galahad=0" > ./install_agenda
RUN echo "let install_cutest=1"    >> ./install_agenda
RUN echo "let install_sifdecode=1" >> ./install_agenda
RUN echo "let install_matlab=0"    >> ./install_agenda
ENV MATLAB_REQUIRED 0

ENTRYPOINT tail -f /dev/null
```

These docker commands will launch/stop the machine (`--name` is useful):

``` shell
docker build -t cutest .  # build the machine
docker run --detach --name mycutest -it cutest  # launch the machine
docker exec -it mycutest bash  # log in
# run ./archdefs/install_optgrove and answer the questions below
export CUTEST_DOCKER=$(docker ps --quiet --filter name=mycutest)
docker stop $CUTEST_DOCKER
docker rm $CUTEST_DOCKER
```

This requires these questions:

* Require CUTEst-Matlab interface? No
* Platform? PC with generic 64-bit processor
* Fortran compiler? GNU gfortran compiler
* C compiler? GNU gcc compiler
* Install SIFDecode? Y
* Install cutest? Y (double version, "D")

I think the list of questions above would be more clear than the shell output included in the documentation (especially since some of the numbers have changed!). Generally, I think the [Install] docs could be tightened up to make it easier to digest. e.g., why waste space on "Manual installation of PyCUTEst", `PYTEST_CACHE`, NumPy or SciPy? Most of those elements are automatically handled or can be delegated to a "Development install" section.

[Install]:https://jfowkes.github.io/pycutest/_build/html/install.html
[doc]:https://jfowkes.github.io/pycutest/_build/html/index.html
[1]:https://jfowkes.github.io/pycutest/_build/html/contributing.html
[c]:https://jfowkes.github.io/pycutest/_build/html/contributing.html
