### 2022-08-08 #4377 @jonjoncardoso
I agree with all the commends made by [stsievert](https://github.com/stsievert), I think these should be considered before acceptance.

Installation was smoother for me, I installed it under the Linux subsystem on Windows and it worked well.

A few things that could be improved:

1. Add a CONTRIBUTING.md. As mentioned by [stsievert](https://github.com/stsievert), it is not exactly clear how one could contribute to this package or what the conventions are (Pull Request for everything? What if I just want to report a bug? Or propose a new feature?)
2. It might be worth to add or point to the [CUTE classification scheme](https://www.cuter.rl.ac.uk/Problems/classification.shtml) somewhere in the paper, the Github repo and the documentation page. Otherwise, one would always have to google what the output of `pycutest.problem_properties` represents; it's better if this is already there and easily accessible in your documentation.
3. I'm not sure how to interpret the problem properties when `n=None` or `m=None`. See for example ARGLALE:
```python
print(pycutest.problem_properties("ARGLALE"))
{'objective': 'N', 'constraints': 'L', 'regular': True, 'degree': 2, 'origin': 'A', 'internal': False,
'n': None, 'm': None}
```
Does it mean there are no variables/constraints that can be changed by the user? If that's so, why can't it just be `n=0` and `m=0`, as in other problem outputs?
