### 2022-08-22 98187f3
Bump to version 1.3 and update release history
