### 2022-08-26 a4464cb
Return 'variable' for problem_properties for variable dimension m or n
