### 2022-08-26 f025cb1
Added CONTRIBUTING guidelines file
