### 2022-08-26 843eaaa
Fix unit tests for variable dimension problems and update find_problems to incorporate new variable-dimension problem_properties convention
