### 2022-08-29 4aa3dfb
Check v is/isn't None for hess and sphess. Closes #27
