### 2022-09-01 32c11cc
Query CI macOS gcc install
