+++ .github/workflows/test_mac.yml
@@ -25,6 +25,8 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
+        ls -lh /usr/local/Cellar/gcc*
+        ls -lh /usr/local/opt/gcc*
         sudo ln -s /usr/local/bin/gfortran-11 /usr/local/bin/gfortran
         sudo mkdir /usr/local/gfortran
         sudo ln -s /usr/local/Cellar/gcc@11/*/lib/gcc/11 /usr/local/gfortran/lib
