### 2022-09-01 cd2f1dd
Switch to latest CI gcc/gfortran
