### 2022-09-01 262ab4d
Revert "Switch to latest CI gcc/gfortran"
This reverts commit cd2f1dd586a198d1761b1b180d511ba17401005c.
