+++ .github/workflows/test_mac.yml
@@ -25,10 +25,9 @@ jobs:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
       run: |
-        ls -lh /usr/local/bin/gcc*
-        ls -lh /usr/local/bin/gfortran*
-        ls -lh /usr/local/Cellar/gcc*
-        ls -lh /usr/local/opt/gcc*
+        sudo ln -s /usr/local/bin/gcc-11 /usr/local/bin/gcc
+        sudo mkdir /usr/local/gcc
+        sudo ln -s /usr/local/Cellar/gcc@11/*/lib/gcc/11 /usr/local/gcc/lib
         sudo ln -s /usr/local/bin/gfortran-11 /usr/local/bin/gfortran
         sudo mkdir /usr/local/gfortran
         sudo ln -s /usr/local/Cellar/gcc@11/*/lib/gcc/11 /usr/local/gfortran/lib
