### 2022-09-01 9afa9d2
Fix macOS CI tests (#29)
