+++ .github/workflows/test_mac.yml
@@ -28,6 +28,7 @@ jobs:
         sudo ln -s /usr/local/bin/gfortran-11 /usr/local/bin/gfortran
         sudo mkdir /usr/local/gfortran
         sudo ln -s /usr/local/Cellar/gcc@11/*/lib/gcc/11 /usr/local/gfortran/lib
+        sudo ln -s /usr/local/Cellar/gcc@11/*/lib/gcc/11 /usr/local/opt/gcc/lib/gcc/11
         brew tap optimizers/cutest &&
         brew install cutest --without-single
         pip install -r requirements.txt --only-binary=numpy,scipy
