### 2022-09-02 fe9051b
Clarify and improve install docs
