### 2022-09-02 4a238ce
Clarify and improve install docs
