+++ README.rst
@@ -13,11 +13,6 @@ PyCUTEst requires the following software to be installed:
 * Python 2.7 or Python 3 (http://www.python.org/)
 * CUTEst (see below)
 
-Additionally, the following python packages should be installed (these will be installed automatically if using *pip*, see `Installing PyCUTEst using pip`_):
-
-* NumPy 1.11 or higher (http://www.numpy.org/)
-* SciPy 0.18 or higher (http://www.scipy.org/)
-
 **Please Note:** Currently PyCUTEst only supports Mac and Linux. For Windows 10 (or later), PyCUTEst can be used through the `Windows Subsystem for Linux <https://docs.microsoft.com/en-us/windows/wsl/>`_, following the Linux installation instructions.
 
 Installing CUTEst on Linux
@@ -54,19 +49,18 @@ Now we are ready to install CUTEst in double precision (requires :code:`gfortran
 
     $ cd /path/to/cutest/cutest/
     $ ${ARCHDEFS}/install_optrove
-    Do you wish to install CUTEst (Y/n)? Y
-    Do you require the CUTEst-Matlab interface (y/N)? N
-    Select platform: 6 # PC with generic 64-bit processor
-    Select operating system: 3 # Linux
-    Would you like to review and modify the system commands (y/N)? N
-    Select fortran compiler: 6 # GNU gfortran compiler
-    Would you like to review and modify the fortran compiler settings (y/N)? N
-    Select C compiler: 8 # GCC
-    Would you like to review and modify the C compiler settings (y/N)? N
-    Would you like to compile SIFDecode (Y/n)? Y
-    Would you like to compile CUTEst (Y/n)? Y
-    CUTEst may be compiled in (S)ingle or (D)ouble precision or (B)oth.
-    Which precision do you require for the installed subset (D/s/b) ? D
+
+This requires answering some questions as follows:
+
+    * Install CUTEst? Yes
+    * Require CUTEst-Matlab interface? No
+    * Platform? PC with generic 64-bit processor
+    * Operating system? Linux
+    * Fortran compiler? GNU gfortran compiler
+    * C compiler? GCC
+    * Compile SIFDecode? Yes
+    * Compile CUTEst? Yes
+    * Precision required? Double
 
 And CUTEst should run from here. To test that the installation works, issue the commands:
 
@@ -98,8 +92,8 @@ Then you can easily install CUTEst:
 
 **Please Note:** *you may see warnings such as* :code:`ld: warning: object file (RANGE.o) was built for newer macOS version (11.5) than being linked (10.15)` *when using PyCUTEst on Mac. To suppress these warnings please set the environment variable* :code:`MACOSX_DEPLOYMENT_TARGET` *to your current macOS version (e.g.* :code:`export MACOSX_DEPLOYMENT_TARGET=11.5` *in this example, you can make this permanent by adding it to your* :code:`~/.bashrc` *file).*
 
-Installing PyCUTEst using pip
------------------------------
+Installing PyCUTEst
+-------------------
 For easy installation, use `pip <http://www.pip-installer.org/>`_:
 
  .. code-block:: bash
@@ -128,33 +122,7 @@ And set an environment variable to tell PyCUTEst about this directory, by adding
 
 If you do not set this environment variable, then PyCUTEst will create a cache folder of compiled problems inside your current working directory.
 
-Manual installation of PyCUTEst
--------------------------------
-Alternatively, you can download the source code from `Github <https://github.com/jfowkes/pycutest>`_ and unpack as follows:
-
- .. code-block:: bash
-
-    $ git clone https://github.com/jfowkes/pycutest
-    $ cd pycutest
-
[…]
