### 2022-09-02 2af58b8
Improve contributing docs section
