### 2022-09-02 959ce6d
Improve contributing docs section
