### 2022-09-05 cffd9b6
Simplify Linux installation
