### 2022-09-05 c455677
Simplify Linux installation
