+++ pycutest/sifdecode_extras.py
@@ -270,11 +270,11 @@ def find_problems(objective=None, constraints=None, regular=None,
 
     See http://www.cuter.rl.ac.uk/Problems/classification.shtml for details on the requirements.
 
-    :param objective: a string containing one or more substrings ('none','constant','linear','quadratic','sum of squares','other') specifying the type of the objective function
-    :param constraints: a string containing one or more substrings ('unconstrained','equality','bound','adjacency','linear','quadratic','other') specifying the type of the constraints
+    :param objective: a string containing one or more substrings (``'none'``, ``'constant'``, ``'linear'``, ``'quadratic'``, ``'sum of squares'``, ``'other'``) specifying the type of the objective function
+    :param constraints: a string containing one or more substrings (``'unconstrained'``, ``'equality'``, ``'bound'``, ``'adjacency'``, ``'linear'``, ``'quadratic'``, ``'other'``) specifying the type of the constraints
     :param regular: a boolean, ``True`` if the problem must be regular, ``False`` if it must be irregular
     :param degree: list of the form ``[min, max]`` specifying the minimum and the maximum number of analytically available derivatives
-    :param origin: a string containing one or more substrings ('academic','modelling','real-world') specifying the origin of the problem
+    :param origin: a string containing one or more substrings (``'academic'``, ``'modelling'``, ``'real-world'``) specifying the origin of the problem
     :param internal: a boolean, ``True`` if the problem must have internal variables, ``False`` if internal variables are not allowed
     :param n: a list of the form ``[min, max]`` specifying the lowest and the highest allowed number of variables
     :param userN: ``True`` if the problems must have user settable number of variables, ``False`` if the number must be hardcoded
