+++ docs/contributing.rst
@@ -4,10 +4,10 @@ Contributing
 Contributing to PyCUTEst
 ------------------------
 
-Contributions to PyCUTEst are always welcome!
+Contributions to PyCUTEst are most welcome.
 
-Reporting Bugs/Feature Suggestions
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Reporting Bugs / Feature Suggestions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 If you find a bug in PyCUTEst, or have a suggestion about how PyCUTEst can be improved, please
 `create a Github issue <https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue>`_.
@@ -17,9 +17,9 @@ rather than creating a new one.
 
 When reporting a bug, it would be helpful if you could include:
 
-* Informative title and clear description
+* An informative title and clear description
 * A code snippet showing how the error occurred
-* Relevant Python output messages
+* Any relevant Python output messages
 * Information about your installation (e.g. operating system, Python version/distribution, PyCUTEst version)
 
 Submitting code changes
@@ -46,4 +46,4 @@ Code of Conduct
 We expect everyone in the PyCUTEst community to show respect to each other and behave appropriately at all times.
 Abusive language, harassment and inappropriate/unprofessional language is not acceptable.
 Instances of unacceptable behavior can be reported to the PyCUTEst team
-([Jari](mailto:jaroslav.fowkes@stfc.ac.uk) and [Lindon](mailto:lindon.roberts@sydney.edu.au)).
+(`Jari <mailto:jaroslav.fowkes@stfc.ac.uk>`_ and `Lindon <mailto:lindon.roberts@sydney.edu.au>`_).
