### 2022-09-06 0314573
Minor clarifications to the docs
