### 2022-09-06 #4377 @jfowkes
@stsievert @jonjoncardoso we have incorporated your suggestions and created a new release with the following major changes:

- we have changed the author order as suggested
- we have added a `CONTRIBUTING.md` and a proper contributing section to the docs
- `problem_properties` and `find_problems()` now use human-readable descriptions as suggested (no need for CUTEst classification codes)
- the Linux CUTEst installation has been simplified to a one-liner without human input (as in the CI)
- similarly for the Dockerfile (no human input is required, has been tested)
- install documentation has been streamlined

Please see jfowkes/pycutest#26 for all the changes.
