### 2022-09-08 1f2391f
Fix typo in README.rst
