### 2022-09-18 #4377 @stsievert
Thanks for the update @jfowkes, the install went really smooth and it's looking much better. I've checked off all but one of the boxes, and have one recommendation:

* *(issue)* Thank you for including the relevant emails, both on the homepage and CONTRIBUTING.md. I'm still not seeing any "clear" guidelines for anyone "seeking support" (though it's not explicit, people seeking support do have a mechanism to obtain support).
* *(recommendation)* I'd use [GitHub templates for pull request/issues][1], which will allow removing the "include this information in your PR/issue" in [CONTRIBUTING.md][2].

[1]:https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository
[2]:https://github.com/jfowkes/pycutest/blob/1f2391f9a0e67bd662c2de3b4466182b1c7b57bf/CONTRIBUTING.md
