+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,20 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behaviour, including a code snippet showing how the error occurred
+and any relevant Python output messages.
+
+**Information about your installation:**
+ - Operating System
+ - Python Version / Distribution
+ - PyCUTEst Version
