+++ .github/pull_request_template.md
@@ -1,0 +1,9 @@
+
+**Describe the change(s)**
+A clear and concise description of the change(s). If you are fixing an issue, please reference the issue number.
+
+**Have you updated the documentation?**
+We use [Sphinx](https://www.sphinx-doc.org/en/master/) to create our documentation which is located in the `docs` subdirectory. You will need to install both the `sphinx` and `sphinx-rtd-theme` packages.
+
+**Have you included relevant unit tests?**
+We use [nose](https://nose.readthedocs.io/en/latest/) to run our unit tests which are located in `pycutest/tests`. We also provide a `tox.ini` for local testing using [tox](https://tox.wiki/en/latest/). 
