### 2022-09-21 f533453
Remove template info from CONTRIBUTING.md
