### 2022-09-21 2d1fd2c
Remove template info from contributing.rst
