+++ docs/contributing.rst
@@ -15,13 +15,6 @@ It would be helpful if you first check the `existing list of open issues <https:
 to see if someone else has reported the same thing - in this case it would be better to comment on the existing issue
 rather than creating a new one.
 
-When reporting a bug, it would be helpful if you could include:
-
-* An informative title and clear description
-* A code snippet showing how the error occurred
-* Any relevant Python output messages
-* Information about your installation (e.g. operating system, Python version/distribution, PyCUTEst version)
-
 Submitting code changes
 ^^^^^^^^^^^^^^^^^^^^^^^
 
@@ -29,17 +22,6 @@ We are happy to accept code contributions from the PyCUTEst community.
 If you have made a change (bugfix, new feature, documentation improvement, etc), then please
 `submit a pull request <https://github.com/jfowkes/pycutest/pulls>`_.
 Github has `more information on pull requests <https://docs.github.com/en/pull-requests>`_.
-Please ensure that:
-
-* Your description clearly explains your changes
-* You have updated the documentation in :code:`docs`.
-  We use `Sphinx <https://www.sphinx-doc.org/en/master/>`_ to create our documentation.
-  You will need to install both the Sphinx and sphinx-rtd-theme packages.
-* You have included relevant unit tests in :code:`pycutest/tests`.
-  We use `nose <https://nose.readthedocs.io/en/latest/>`_ to run our unit tests
-  but we do also provide a :code:`tox.ini` for local testing using `tox <https://tox.wiki/en/latest/>`_.
-
-If you are fixing an issue, please reference the issue number in the pull request.
 
 Code of Conduct
 ^^^^^^^^^^^^^^^
