+++ docs/support.rst
@@ -1,0 +1,9 @@
+Support
+=======
+
+Support for PyCUTEst
+--------------------
+
+If you require any support with using PyCUTEst, please feel free to either email the PyCUTEst team
+(`Jari <mailto:jaroslav.fowkes@stfc.ac.uk>`_ and `Lindon <mailto:lindon.roberts@sydney.edu.au>`_)
+or `open a GitHub issue <https://github.com/jfowkes/pycutest/issues>`_ with your query.
