+++ docs/contributing.rst
@@ -10,8 +10,8 @@ Reporting Bugs / Feature Suggestions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 If you find a bug in PyCUTEst, or have a suggestion about how PyCUTEst can be improved, please
-`create a Github issue <https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue>`_.
-It would be helpful if you first check the `existing list of open issues <https://github.com/jfowkes/pycutest/issues>`_
+`create a GitHub issue <https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue>`_.
+It would be helpful if you could first check the `existing list of open issues <https://github.com/jfowkes/pycutest/issues>`_
 to see if someone else has reported the same thing - in this case it would be better to comment on the existing issue
 rather than creating a new one.
 
@@ -21,7 +21,7 @@ Submitting code changes
 We are happy to accept code contributions from the PyCUTEst community.
 If you have made a change (bugfix, new feature, documentation improvement, etc), then please
 `submit a pull request <https://github.com/jfowkes/pycutest/pulls>`_.
-Github has `more information on pull requests <https://docs.github.com/en/pull-requests>`_.
+GitHub has `more information on pull requests <https://docs.github.com/en/pull-requests>`_.
 
 Code of Conduct
 ^^^^^^^^^^^^^^^
