### 2022-09-27 #4377 @jfowkes
Many thanks @stsievert we've now merged both the outstanding issue and your suggestion into the docs.
