+++ requirements.txt
@@ -1,4 +1,2 @@
-numpy>=1.11; python_version>="3"
-scipy>=0.18; python_version>="3"
-numpy>=1.11,<1.17; python_version<"3"
-scipy>=0.18,<1.3; python_version<"3"
+numpy>=1.11
+scipy>=0.18
