### 2022-09-27 b937865
C interface modernisation
