### 2022-09-27 98c4a82
C interface modernisation
