+++ pycutest/c_interface.py
@@ -1,21 +1,14 @@
 """
-C source code of the CUTEst interface
-
-Note that for Python 2 and 3 compatibility, we need to make some modifications to itf_c_source (see end of this file).
+C source code of the CUTEst problem interface
 """
 __all__ = ['itf_c_source']
 
 # Because we dont want backslashes to be interpreted as escape characters, the string must be a raw string.
 itf_c_source = r"""
-/* CUTEst interface to Python and NumPy */
+/* CUTEst problem interface to Python and NumPy */
 /* (c)2011 Arpad Buermen */
-/* (c)2018 Jaroslav Fowkes, Lindon Roberts */
+/* (c)2022 Jaroslav Fowkes, Lindon Roberts */
 /* Licensed under GNU GPL V3 */
-
-/* Note that in Windows we do not use Debug compile because we don't have the debug version
-   of Python libraries and interpreter. We use Release version instead where optimizations
-   are disabled. Such a Release version can be debugged.
- */
 
 /* Unused CUTEst tools - sparse finite element matrices and banded matrices
      cdimse
@@ -35,8 +28,7 @@ itf_c_source = r"""
      cscfg		... obsolete
      cscifg		... obsolete
 */
-
-#define NPY_NO_DEPRECATED_API NPY_1_8_API_VERSION
+#define NPY_NO_DEPRECATED_API NPY_1_20_API_VERSION
 
 #include <Python.h>
 #include <cutest.h>
@@ -44,23 +36,8 @@ itf_c_source = r"""
 #include <math.h>
 #include <stdio.h>
 
-/* Debug switch - uncomment to enable debug messages */
-/* #define PYDEBUG */
 
-/* Debug file */
-#define df stdout
-
-#ifndef WIN32
-#define __declspec(a)
-#endif
-
-/* Safeguard against C++ symbol mangling */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-
-/* Prototypes */
+/* Module function prototypes */
 static PyObject *cutest__dims(PyObject *self, PyObject *args);
 static PyObject *cutest__setup(PyObject *self, PyObject *args);
 static PyObject *cutest__varnames(PyObject *self, PyObject *args);
@@ -81,7 +58,7 @@ static PyObject *cutest__isphess(PyObject *self, PyObject *args);
 static PyObject *cutest__gradsphess(PyObject *self, PyObject *args);
 static PyObject *cutest_report(PyObject *self, PyObject *args);
 
-/* Persistent data */
+/* Module global variables */
 #define STR_LEN 10
 static npy_int status = 0;    /* output status */
 static npy_int CUTEst_nvar = 0;		/* number of variables */
@@ -100,14 +77,12 @@ static char  fName[] = "OUTSDIF.d"; 	/* Data file name */
 static logical somethingFalse = FALSE_, somethingTrue = TRUE_;
 
 
-/* Helper functions */
+/* Module helper functions */
 
 /* Open data file, return 0 on error. */
 int open_datafile(void) {
     npy_int  ioErr;					/* Exit flag from OPEN and CLOSE */
-#ifdef PYDEBUG
-    fprintf(df, "PyCUTEst: opening data file\n");
-#endif
+
     ioErr = 0;
     if (! dataFileOpen)
[…]
