+++ pycutest/problem_class.py
@@ -2,9 +2,6 @@
 A class to store problem info, where we can set up the interface exactly how we wish
 """
 
-# Ensure compatibility with Python 2
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 import numpy as np
 
 __all__ = ['CUTEstProblem']
