### 2022-09-29 383356c
Add install note for Anaconda users
