### 2022-10-23 #4377 @diehlpk
@jfowkes

I have some minor editorial comments on the paper:

- `With over 12,000 downloads we firmly believe `, please add something like `at the time this paper was written` or `by 2022` to make it clear when this amount was measured.
- `PyCUTEst has had over 12,000 downloads, and we believe is well on ` please see above.

Please address these timely, so we can publish the paper next week.

After you addressed these comments, please do the following:

- [ ] Generate a new release including all changes done during the review and post the number here.
- [ ] Please upload that release to Zenodo or FigShare and post the DOI here. Note that the title and authors of the archive have to be consitent with the paper.
