### 2022-10-24 bd8864a
Update download stats for paper
