### 2022-10-24 #4377 @jfowkes
@diehlpk thank you, I have addressed the minor editorial comments in the paper.

We have also released a new version (Version 1.4) with all the changes.

This has been uploaded to Zenodo under the following DOI:
[10.5281/zenodo.7244291](https://doi.org/10.5281/zenodo.7244291)
