### 2022-10-24 #4377 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
