### 2022-10-25 #4377 @danielskatz
@jfowkes - I'm the track editor who will accept and publish this paper.  I have a few minor changes that I've suggested in https://github.com/jfowkes/pycutest/pull/35 - please merge this or let me know what you disagree with, then we can complete the process.
