### 2022-10-25 #4377 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
